updated readme, config, command for removing 2fa
