+++ components/2FABundle/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,11 @@
+Thanks for your pull request! We love contributions.
+
+However, this repository is what we call a "subtree split": a read-only copy of one directory of the main repository. It is used by Composer to allow developers to depend on specific bundles.
+
+If you want to contribute, you should instead open a pull request on the main repository:
+
+https://github.com/Novactive/Nova-eZPlatform-Bundles
+
+Thank you for your contribution!
+
+PS: if you haven't already, please add tests.
