+++ components/2FABundle/.github/issue_template.md
@@ -1,0 +1,9 @@
+Thanks for reporting an issue! We love feedback.
+
+However, this repository is what we call a "subtree split": a read-only copy of one directory of the main repository. It is used by Composer to allow developers to depend on specific bundles.
+
+If you want to report or contribute, you should instead open your issue on the main repository:
+
+https://github.com/Novactive/Nova-eZPlatform-Bundles
+
+Thank you for your contribution!
