+++ components/2FABundle/.github/workflows/auto-close-issue.yaml
@@ -1,0 +1,21 @@
+on:
+    issues:
+        types: [opened, edited]
+
+jobs:
+    autoclose:
+        runs-on: ubuntu-latest
+        steps:
+             - name: Close Issue
+               uses: peter-evans/close-issue@v1
+               with:
+                 comment: |
+                     Thanks for reporting an issue! We love feedback.
+
+                     However, this repository is what we call a "subtree split": a read-only copy of one directory of the main repository. It is used by Composer to allow developers to depend on specific bundles.
+
+                     If you want to report or contribute, you should instead open your issue on the main repository:
+
+                     https://github.com/Novactive/Nova-eZPlatform-Bundles
+
+                     Thank you for your contribution!
