+++ components/2FABundle/.github/workflows/auto-close-pr.yaml
@@ -1,0 +1,23 @@
+on:
+    pull_request:
+        types: [opened,edited,reopened]
+
+jobs:
+    autoclose:
+        runs-on: ubuntu-latest
+        steps:
+             - name: Close Pull Request
+               uses: peter-evans/close-pull@v1
+               with:
+                 comment: |
+                        Thanks for your pull request! We love contributions.
+
+                        However, this repository is what we call a "subtree split": a read-only copy of one directory of the main repository. It is used by Composer to allow developers to depend on specific bundles.
+
+                        If you want to contribute, you should instead open a pull request on the main repository:
+
+                        https://github.com/Novactive/Nova-eZPlatform-Bundles
+
+                        Thank you for your contribution!
+
+                        PS: if you haven't already, please add tests.
