+++ components/AlgoliaSearchEngine/README.md
@@ -7,7 +7,7 @@ It is used by Composer to allow developers to depend on specific bundles.
 
 If you want to report or contribute, you should instead open your issue on the main repository: https://github.com/Novactive/Nova-eZPlatform-Bundles
 
-Documentation is available in this repository via `.md` files but also packaged here: https://novactive.github.io/Nova-eZPlatform-Bundles/master/NovaeZAlgoliaSearchEngine/README.md.html
+Documentation is available in this repository via `.md` files but also packaged here: https://novactive.github.io/Nova-eZPlatform-Bundles/master/AlgoliaSearchEngine/README.md.html
 
 ----
